requirements.txt
pywebview>=5.0
pywxdump>=3.1
playwright>=1.40.0
pycryptodome>=3.19.0
Pillow>=10.0.0
